Revert "Revert "Merge StoragePolicy down into ContentPolicy, and always use this""
